repackages fetcher consumer
